ex resolvidos e mais

+++ ex4.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
 #-*- coding:utf-8 -*-
 
-n1 = int(input())
-n2 = int(input())
-n3 = int(input())
-n4 = int(input())
+n1 = float(input())
+n2 = float(input())
+n3 = float(input())
+n4 = float(input())
 s = n1 + n2 + n3 + n4
 print("A média final é:", s / 4)
